requirements.txt
black
pandas
scikit-learn
matplotlib
skops
gradio
pillow
torch
numpy
onnx
torchvision
streamlit
onnxruntime
mlflow
mlflow[extras]
